Fix formatting of ONIE's install logs
The logs were added outside of a code block in order to highlight two
specific lines; and blank lines had to be added between each line to
prevent the backend from concatenating all lines when rendering the
HTML. As a result, the logs are not exactly pretty to read.

Instead, we can move them to a code block and still benefit from
highlighting, by using the dedicated option for the block and
emphasasing the two relevant lines.

Signed-off-by: Quentin Monnet <qmo@qmon.net>