Clean-up style, formatting, typos
This is a major clean-up for documentation style, formatting, typos.
Changes include:

- Style: I applied a number of generic recommendations that I've read in
  the past in style guides for documentation. For example:

    - Generally improve style consistency on case, formatting,
      introducing commands or flags, or a number of other small items
    - Avoid the use of the first person ("We/our"), as it is unclear who
      "we" refers to. Sometimes it seems to be the author and reader,
      sometimes it seems to be Hedgehog; replace accordingly.
    - Use plain, direct language and shorten some sentences
    - Use specific wording rather than vague wording. Avoid "it" or
      "this" when it is not entirely clear what they refer to.
    - Use present tense
    - Avoid abbreviations, including Latin abbreviations, especially
      "and etc" which is not correct and got me to start fixing in the
      first place
    - Use upper case for initialisms (for example, "VXLAN")

- Formatting:

    - Some text rewrapping on 120 characters
    - Add blank lines below headings, above code blocks

- Various typo or grammar fixes:

    - Including a number of edits related to the use of
      articles ("a"/"the"/"")

I cannot pretend I have managed to apply all the principles above
everywhere in the docs, but I did my best to improve the text where I
could. I certainly missed some occurrences to improve.

Pages under docs/reference (and their typos) are left untouched, as I
understand their contents is generated from another source.

Other than styling, the technical contents, the nature of the
directives, the meaning of the text should not be modified in this
commit. If instructions do change, this is a mistake.

Signed-off-by: Quentin Monnet <qmo@qmon.net>