Consent related forms (#23)
Turn on the cookie banner - 
We might not NEED it for California stuff -
[here](https://cppa.ca.gov/faq.html#faq_cppa_4_body)
According to the material-mkdocs GDPR says we need to be able to cookie
consent at any time. This has been added. We are relying on
material-mkdocs to have done the plumbing to actually turn off the
tracking when the user says so.

---------

Co-authored-by: Sergei Lukianov <me@slukjanov.name>