Upgrade instructions: Add tabs for Linux/macOS instructions
Introduce the use of content tabs in the docs, because it looks nice and
fancy. In fact we already had tabs in the contribution guide for docs,
but that one was likely a copy-paste from Mkdocs documentation.

Link content tabs (although for now we only have one tab group):
https://squidfunk.github.io/mkdocs-material/reference/content-tabs/#linked-content-tabs

Fix the case on “macOS”.

Signed-off-by: Quentin Monnet <qmo@qmon.net>